Add a function to create a problem.
